Move SSLPreHandler to its own file
To avoid having multiple top-level classes in one file
