+++ testing/internal-integ-testing/src/main/groovy/org/gradle/test/fixtures/server/http/HttpServerFixture.groovy
@@ -18,7 +18,6 @@ package org.gradle.test.fixtures.server.http
 
 import groovy.transform.CompileStatic
 import org.eclipse.jetty.http.HttpVersion
-import org.eclipse.jetty.server.Connector
 import org.eclipse.jetty.server.Handler
 import org.eclipse.jetty.server.HttpConfiguration
 import org.eclipse.jetty.server.HttpConnectionFactory
@@ -216,22 +215,5 @@ trait HttpServerFixture {
         connector.stop()
         connector.close()
         server?.removeConnector(connector)
-    }
-}
-
-@CompileStatic
-class SslPreHandler implements HttpConfiguration.Customizer {
-
-    private final List<Consumer<Request>> consumers = []
-
-    void registerCustomizer(Consumer<Request> consumer) {
-        consumers << consumer
-    }
-
-    @Override
-    void customize(Connector connector, HttpConfiguration channelConfig, Request request) {
-        consumers.each {
-            it.accept(request)
-        }
     }
 }
